refactor: move CMerkleTx into the wallet, GetTransaction into validation
Move the CMerkleTx class from main.h into wallet/wallet.h and its method
bodies (SetMerkleBranch, GetDepthInMainChain[INTERNAL], GetBlocksToMaturity,
AcceptToMemoryPool) plus CWalletTx::AcceptWalletTransaction from main.cpp
into wallet/wallet.cpp -- only CWalletTx derives from CMerkleTx, mirroring
Bitcoin's relocation of the class into the wallet (issue #3125, workstream
C3). The free GetTransaction(hash, tx, hashBlock) chain/mempool lookup has
no wallet dependency and moves to validation.{h,cpp} instead. Pure code
move, no behavior change.

main.h keeps compiling for its ~98 includers: GetTransaction's declaration
now comes transitively from validation.h (already included by main.h), and
the only CMerkleTx consumer that was not already including wallet/wallet.h
(rpc/blockchain.cpp, for its txGen locals) now does.

main.cpp -139 lines, main.h -53 lines.
